+++ assertk/src/commonMain/kotlin/assertk/assertions/array.kt
@@ -182,13 +182,13 @@ fun Assert<Array<*>>.containsExactly(vararg elements: Any?) = given { actual ->
  * Asserts the collection contains at least one instance of a given type.
  *
  * ```
- * assertThat(arrayOf<Any>("one", "two", 1)).containsInstanceOf<String>().each {
+ * assertThat(arrayOf<Any>("one", "two", 1)).havingInstanceOf<String>().each {
  *   it.hasLength(3)
  * }
  * ```
  */
-inline fun <reified T> Assert<Array<*>>.containsInstanceOf(): Assert<List<T>> {
-    return transform("contains subtype of ${T::class}") { actual ->
+inline fun <reified T> Assert<Array<*>>.havingInstancesOf(): Assert<List<T>> {
+    return transform("contains instances of ${T::class}") { actual ->
         actual.filterIsInstance<T>().also {
             if (it.isEmpty()) expected("to contain at least one instance of ${T::class} but was ${actual.asList()}")
         }
@@ -199,10 +199,10 @@ inline fun <reified T> Assert<Array<*>>.containsInstanceOf(): Assert<List<T>> {
  * Asserts the collection does not contain an instance of a given type.
  *
  * ```
- * assertThat(arrayOf<Any>("one", "two", 1)).doesNotContainInstanceOf<Double>()
+ * assertThat(arrayOf<Any>("one", "two", 1)).havingInstancesOf<Double>()
  * ```
  */
-inline fun <reified T> Assert<Array<*>>.doesNotContainInstanceOf() = given { actual ->
+inline fun <reified T> Assert<Array<*>>.notHavingInstancesOf() = given { actual ->
     if (actual.any { it is T }) expected("to not contain instances of ${T::class} but was ${actual.asList()}")
 }